chore: move UserInfo.kt
