+++ app/src/main/java/com/sunshine/android/ui/feature/home/HomeScreen.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.sunshine.android.R
-import com.sunshine.android.data.UserInfo
+import com.sunshine.android.data.model.UserInfo
 import com.sunshine.android.ui.theme.Brown
 import com.sunshine.android.ui.theme.Red
 import com.sunshine.android.ui.theme.Yellow
